+++ src/wrapper.ts
@@ -99,7 +99,8 @@ export const NATIVE = {
 
     // filter out all options that would crash native
     const filteredOptions = FilterNativeOptions(options);
-    return SentryCapacitor.initNativeSdk({ options: filteredOptions });
+    const nativeIsReady = await SentryCapacitor.initNativeSdk({ options: filteredOptions });
+    return nativeIsReady;
   },
 
   /**